+++ release/sil/sil_devanagari_romanized/HISTORY.md
@@ -3,7 +3,7 @@ Devanagari Romanized (SIL) Change History
 
 1.0 (2020-01-24)
 -------------------------
-* Initial release
+* Initial release.
 
 1.5 (2022-01-10)
 -------------------------
@@ -19,3 +19,20 @@ Devanagari Romanized (SIL) Change History
 -------------------------
 * Devanagari Letter Vocalic- R /U+090B/ (LCtl-LAlt) is fixed now.
 * Devanagari Double Danda > /U+0965/ Shift + | is fixed now.
+
+1.7 (2022-03-03)
+-------------------------
+/*New Addition*/
+
+*Devanagari Letter Vocalic- L / U+090C/
+*Devanagari Letter Candra E /U+090D/
+*Devanagari Letter Short E/U+090E/
+*Devanagari Letter Candra O/U+0911/
+*Devanagari Letter Short O/U+0912/
+*Devanagari Letter Vocalic Rr/U+0960/
+*Devanagari Letter Vocalic- Ll/U+0961/
+
+/*Removed*/
+*The default 'x' behavior for post-h is non-joining for Kangri has been removed.
+
+*Minor Changes in phone keyboard layout.